[p2] add flatbuffer
